platform/x86: touchscreen_dmi: Match on ACPI HID instead of ACPI companion-dev-name
The ACPI-companion device's dev_name() (which gets used as
i2c_client->name for I2C-devices) consists of the ACPI HID to a ":##"
suffix, which typically is ":00" but if there is more then one ACPI
fwnode with the same HID it may be ":01" etc.

On some tablets there are multiple ACPI fwnode-s describing the
touchscreen, one for each I2C bus the device may be connected to.
In this case we get ACPI-companion devices with e.g. "GDIX1001:00"
and "GDIX1001:01" as dev_name(). This makes matching on the dev_name()
as we have been doing inconvinient if we want to share a ts_dmi_data
struct between device's which need the same set of device-properties,
bit have their touchscreen on different busses leading to different
":##" suffixes.

Since there always is only one touchscreen, there is no reason to
include the suffix in the matching.

Change the code to match on ACPI HID instead of ACPI-companion device's
dev_name() so that it is more flexible.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>