mfd: tps68470: Update header macros
This commit contains a number of updates to header macros for the
tps68470 header:

The value for TPS68470_OSC_EXT_CAP_DEFAULT is updated to set 10pF of
capacitance instead of 0.

The value for TPS68470_DRV_STR_2MA is corrected - 0x00 actually sets
drive strength of 1mA for the two clock outputs, not 2mA.

The value for TPS68470_PLL_OUTPUT_ENABLE is corrected - BIT(4) actually
sets a reserved bit in that register. The correct value is 0x02 (the
setting is two bits wide, so BIT() seems the wrong choice anyway).

Finally, the representation of the value for TPS68470_OUTPUT_B_SHIFT is
switched from BIT(1) to 0x02 - again because it seems like the wrong
choice to use BIT() to represent a value potentially wider than one bit.

Signed-off-by: Daniel Scally <djrscally@gmail.com>