regulator: tps68470 Drop enum tps68470_regulators
This enum will be part of the new include/linux/platform_data/tps68470.h
file, which will be added in a separate patch which will be a pre-requisite
for the patch adding the new tps68470-regulator driver.

Note this should be squashed into the initial driver submission patch,
before submitting upstream.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>