regulator: tps68470: Use a normal for-loop to register the regulators
Use a normal for-loop to register the regulators, instead of the weird
reverse-order while loop which makes it hard to see how much regulators
are being registered because that is hidden in the initialization of 'i'
higher up.

Note this should be squashed into the initial driver submission patch,
before submitting upstream.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>