platform/x86: intel_skl_int3472: Add get_sensor_adev_and_name() helper
The intel_skl_int3472_discrete.c code is not the only code which needs
to lookup the acpi_device and device-name for the sensor for which the
INT3472 ACPI-device is a GPIO/clk/regulator provider.

The intel_skl_int3472_tps68470.c code also needs this functionality,
so factor this out into a new get_sensor_adev_and_name() helper.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>