platform/x86: int3472: Pass tps68470_regulator_platform_data to the tps68470-regulator MFD-cell
Pass tps68470_regulator_platform_data to the tps68470-regulator
MFD-cell, specifying the voltages of the various regulators and
tying the regulators to the sensor supplies so that sensors which use
the TPS68470 can find their regulators.

Since the voltages and supply connections are board-specific, this
introduces a DMI matches int3472_tps68470_board_data struct which
contains the necessary per-board info.

This per-board info also includes GPIO lookup information for the
sensor IO lines which may be connected to the tps68470 GPIOs.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v6:
- Also list the VCM module as VSIO regulator consumer
- Add constraints for VIO, it MUST have the same voltage configured as
  VSIO and mark it as always-on (it is always on there is no on/off
  control for it in the PMIC, as the PMIC uses it internally)

Changes in v5:
- Also pass regulator_init_data for the regulator used for the VCM
  (Voice Coil Motor) used for the focus of the back-camera lens