media: ipu3-cio2: Store cio2_bridge pointer in struct cio2_device
This is a preparation patch for adding support for instantiating
i2c-clients for VCMs described in SSDB ACPI tables.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Note there seems to be a pre-existing problem where there is no teardown
of the bridge? It looks like we need a cio2_bridge_cleanup() function,
to be called from cio2_pci_remove() ?