platform/x86: xiaomi-mipad2: New driver for Xiaomi Mi Pad 2 tablets
The ACPI tables on the Xiaomi Mi Pad 2 hide a number of devices from Linux,
such as the fuel-gauge IC used for battery monitoring.

This adds a xiaomi-mipad2 "board-file" which manually instantiates
device-objects for the missing devices.

When build as a module this will only ever get auto-loaded on Xiaomi
Mi Pad 2 tablets, allowing us to support these tablets in generic X86
kernels at the cost of just a small amount of diskspace for the module.

As discussed previously this is the least ugly option to get these
devices to fully work and to do so without adding any extra code to
the main kernel image (vmlinuz) when built as a module.

Link: https://lore.kernel.org/platform-driver-x86/20211031162428.22368-1-hdegoede@redhat.com/
Signed-off-by: Hans de Goede <hdegoede@redhat.com>